+++ scripts/esbuild.mjs
@@ -7,6 +7,16 @@ await build({
     format: "cjs",
     target: "es2022",
     outfile: "out/web/extension.js",
-    external: ["vscode"],
+    external: [
+        "@xhayper/discord-rpc",
+        "vscode",
+        "net",
+        "events",
+        "tls",
+        "fs",
+        "path",
+        "buffer",
+        "stream"
+    ],
     sourcemap: true,
 }).catch(() => process.exit(1))