+++ src/diagnostics.ts
@@ -5,8 +5,8 @@ import {
 	FunctionPrefixRegex,
 	FunctionScanRegex,
 	getExtensionConfig,
-	isComment,
 	isEscaped,
+	isIgnored,
 	Languages,
 	locateCodeBlock,
 	splitArgs,
@@ -38,7 +38,7 @@ export async function validateDocument(
 	while ((match = ScanRegex.exec(text))) {
 		const index = match.index
 		const start = document.positionAt(index)
-		if (!locateCodeBlock(document, start) || isEscaped(text, index) || isComment(text, index)) continue
+		if (!locateCodeBlock(document, start) || isEscaped(text, index) || isIgnored(text, index)) continue
 
 		const full = match[0]
 		const hasOpening = full.endsWith("[")